PHPLintBear: Move to PHP package
